parser/tests/tree_construction/basic.dat

#data
<style><!--</style><meta><script>--><link></script>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <style>
|       "<!--"
|     <meta>
|     <script>
|       "--><link>"
|   <body>
